Do not use pre-annotated prodigal files
We found a bug in which using the pprodigal annotated files doesn't work, bin scores cannot be calculated despite having single copy genes founds in the faa file. To fix this, we let DASTool annotate the proteins. This doesn't add too much time to this step of the workflow but ensures that it goes to the next step (checkm2 and gtdbtk). We still use the default bin_score of >0.5 for refining.