If-Else statement to check if refined bins output folder is empty.
Case: If we re-run the refining step (e.g. due to a hold error, manually stopping the DAG, rerunning previous parts of the DAG), the output refine folder in `/staging` might not be empty. It might contain .fa files from previous runs.

Problem: This makes it so that there are more .fa files than expected in the refined bins folder, which is then used in downstream steps, like CheckM2 and GTDB-tk.

Solution: We added an if-else statement to check if the refine bins folder is empty. if it's empty then we proceed with dastool. If it is not empty, the refined bins folder will be clear (removed files) before proceeding.